workaround shipping price data type

+++ features/transaction/atc_common/src/main/java/com/tokopedia/atc_common/data/model/request/AddToCartOcsRequestParams.kt
@@ -38,7 +38,7 @@ data class AddToCartOcsRequestParams(
         var isTradeIn: Boolean = false,
 
         @SerializedName("shipping_price")
-        var shippingPrice: Int = 0,
+        var shippingPrice: Double = 0.0,
 
         // analytics data
         var productName: String = "",
